completed home and about pages
